refactor some differ responsibilities into classes
